+++ ktor/client/src/commonMain/kotlin/dev/inmo/micro_utils/ktor/client/TemporalUpload.kt
@@ -1,0 +1,19 @@
+package dev.inmo.micro_utils.ktor.client
+
+import dev.inmo.micro_utils.common.MPPFile
+import dev.inmo.micro_utils.ktor.common.*
+import io.ktor.client.HttpClient
+
+expect suspend fun HttpClient.tempUpload(
+    fullTempUploadDraftPath: String,
+    file: MPPFile,
+    onUpload: (uploaded: Long, count: Long) -> Unit = { _, _ -> }
+): TemporalFileId
+
+suspend fun UnifiedRequester.tempUpload(
+    fullTempUploadDraftPath: String,
+    file: MPPFile,
+    onUpload: (uploaded: Long, count: Long) -> Unit = { _, _ -> }
+): TemporalFileId = client.tempUpload(
+    fullTempUploadDraftPath, file, onUpload
+)