+++ ktor/common/src/commonMain/kotlin/dev/inmo/micro_utils/ktor/common/TemporalFiles.kt
@@ -1,0 +1,8 @@
+package dev.inmo.micro_utils.ktor.common
+
+import kotlin.jvm.JvmInline
+
+const val DefaultTemporalFilesSubPath = "temp_upload"
+
+@JvmInline
+value class TemporalFileId(val string: String)